feat: add tsconfig and package structure
