--- QuizSets.ts
@@ -1,215 +1,0 @@
-export const questionSetOne = [
-  {
-    question: "React is a _________.",
-    answers: [
-      "Web development Framework",
-      "JavaScript Library",
-      "jQuery",
-      "Web Server",
-    ],
-    correctAnswerIndex: 1,
-  },
-  {
-    question: "Which ReactJS function renders HTML to the web page?",
-    answers: [
-      "render()",
-      "ReactDOM.render()",
-      "renders()",
-      "ReactDOM.renders()",
-    ],
-    correctAnswerIndex: 1,
-  },
-  {
-    question: "JSX stands for _____.",
-    answers: ["JSON", "JSON XML", "JavaScript XML", "JavaScript and AngularJS"],
-    correctAnswerIndex: 2,
-  },
-  {
-    question: " JSX allows us to write _____.?",
-    answers: [
-      "jQuery in React",
-      "Angular Code in React",
-      "MySQL in React",
-      "HTML in React",
-    ],
-    correctAnswerIndex: 3,
-  },
-  {
-    question: "What are Props?",
-    answers: [
-      "Props are arguments passed into React components",
-      "Props are functions in the ReactJS",
-      "Props are used to returns multiple values from the function",
-      "All of the above",
-    ],
-    correctAnswerIndex: 0,
-  },
-  {
-    question: "Which method is used to generate lists?",
-    answers: ["map()", "generate()", "new()", "maps()"],
-    correctAnswerIndex: 0,
-  },
-];
-
-export const questionSetTwo = [
-  {
-    question: "What types of components do we have in ReactJS?",
-    answers: [
-      "Variable and Const components",
-      "Class and Function components",
-      "Method and Function components",
-      "Array and Object components",
-    ],
-    correctAnswerIndex: 1,
-  },
-  {
-    question: "Which statement is required to define a class component?",
-    answers: [
-      "extends React.Components",
-      "imports React.Components",
-      "extends React.Component",
-      "imports React.Component",
-    ],
-    correctAnswerIndex: 2,
-  },
-  {
-    question: "Can components be passed as props?",
-    answers: ["Yes", "No", "Maybe"],
-    correctAnswerIndex: 0,
-  },
-  {
-    question: "In ReactJS, what is State?",
-    answers: [
-      "It's a temporary storage of the elements",
-      "It's a state of the execution of the ReactJS application",
-      "It's an internal storage of the components",
-      "All of the above",
-    ],
-    correctAnswerIndex: 2,
-  },
-  {
-    question: "How many elements can a valid react component return?",
-    answers: [
-      "1 Element",
-      "More than 1 element",
-      "React does not return element",
-      "None of the mentioned",
-    ],
-    correctAnswerIndex: 0,
-  },
-  {
-    question: "Which of the following methods is not a part of ReactDOM?",
-    answers: [
-      "ReactDOM.hydrate()",
-      "ReactDOM.destroy()",
-      "ReactDOM.createPortal()",
-      "All of the mentioned",
-    ],
-    correctAnswerIndex: 1,
-  },
-];
-
-export const questionSetThree = [
-  {
-    question:
-      "What of the following is used in React.js to increase performance?",
-    answers: [
-      "Original DOM",
-      "Virtual DOM",
-      "Both A and B.",
-      "None of the above.",
-    ],
-    correctAnswerIndex: 1,
-  },
-  {
-    question:
-      "Which of the following keyword is used to create a class inheritance?",
-    answers: ["Create", "Inherits", "Extends", "This"],
-    correctAnswerIndex: 2,
-  },
-  {
-    question:
-      "What is the declarative way to render a dynamic list of components based on values in an array?",
-    answers: [
-      "Using the reduce array method",
-      "Using the <Each /> component",
-      "Using the Array.map() method",
-      "With a for/while loop",
-    ],
-    correctAnswerIndex: 2,
-  },
-  {
-    question: "What are the two ways to handle data in React?",
-    answers: [
-      "State & Props",
-      "Services & Components",
-      "State & Services",
-      "State & Component",
-    ],
-    correctAnswerIndex: 0,
-  },
-  {
-    question: "Does React.js create a VIRTUAL DOM in the memory?",
-    answers: ["Yes", "No", "Maybe"],
-    correctAnswerIndex: 0,
-  },
-  {
-    question: "What is Reconciliation?",
-    answers: [
-      "The process through which React deletes the DOM",
-      "The process through which React updates and deletes the component",
-      "It is a process to set the state",
-      "The process through which React updates the DOM",
-    ],
-    correctAnswerIndex: 2,
-  },
-];
-
-export const questionSetFour = [
-  {
-    question: "Who developed React.js?",
-    answers: ["Apple", "Facebook", "Twitter", "Google"],
-    correctAnswerIndex: 1,
-  },
-  {
-    question:
-      "Life cycle of components methods falls under following categories?",
-    answers: [
-      "Mounting, Unmounting",
-      "Mounting, Updating",
-      "Updating, Unmounting",
-      "Mounting, Updating, Unmounting",
-    ],
-    correctAnswerIndex: 3,
-  },
-  {
-    question: "React.js covers only the view layer of the app.",
-    answers: ["Yes", "No", "Maybe"],
-    correctAnswerIndex: 0,
-  },
-  {
-    question: "What is the DOM?",
-    answers: [
-      "Document Object Model",
-      "Data Object Model",
-      "Data Option Model",
-      "Documentation Object Model",
-    ],
-    correctAnswerIndex: 0,
-  },
-  {
-    question: "React components can be?",
-    answers: [
-      "UseState, Stateless",
-      "Statefull, Stateless",
-      "Unstatefull Statefull",
-      "None of these",
-    ],
-    correctAnswerIndex: 1,
-  },
-  {
-    question: "Can we write a comment in SVG file?",
-    answers: ["Yes", "No", "Maybe"],
-    correctAnswerIndex: 0,
-  },
-];
