+++ package.json
@@ -1,8 +1,12 @@
 {
   "name": "clue-hunt-next",
   "version": "0.1.0",
-  "homepage": "https://sale4319.github.io/",
   "private": true,
+  "type": "module",
+  "workspaces": [
+    "src/shared/components/*",
+    "src/shared/utils/*"
+  ],
   "scripts": {
     "dev": "next dev",
     "build": "next build",
@@ -11,17 +15,16 @@
     "lint:fix": "next lint --fix"
   },
   "dependencies": {
-    "chn-ui": "^0.0.5",
-    "distributed-ui": "^0.3.0",
+    "chn-ui": "^0.1.2",
     "next": "15.0.3",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "server-only": "^0.0.1"
   },
   "devDependencies": {
     "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
     "eslint": "^8",
     "eslint-config-next": "15.0.3",
     "typescript": "^5"
