+++ src/app/level/(routes)/start/page.tsx
@@ -1,5 +1,5 @@
-import "server-only";
 import Start from "./_components/Start";
+import "server-only";
 
 import { Page } from "@/shared/components/Page";
