+++ src/app/level/(routes)/two/page.tsx
@@ -1,7 +1,7 @@
+import { DefaultButton } from "@app/default-button";
+import { Page } from "@app/page-component";
+import { getRoute } from "@app/utils";
 import "server-only";
-import { DefaultButton } from "@/shared/components/DefaultButton";
-import { getRoute } from "@/shared/utils";
-import { Page } from "@/shared/components/Page/";
 
 export default function page() {
   return (
