+++ src/shared/components/DefaultButton/src/DefaultButton.module.css
@@ -18,7 +18,9 @@
 
 .close:hover,
 .close:focus {
-  box-shadow: inset -3.5em 0 0 0 var(--hover), inset 3.5em 0 0 0 var(--hover);
+  box-shadow:
+    inset -3.5em 0 0 0 var(--hover),
+    inset 3.5em 0 0 0 var(--hover);
 }
 .raise:hover,
 .raise:focus {
@@ -35,11 +37,15 @@
   box-shadow: inset 8em 0 0 0 var(--hover);
 }
 .offset {
-  box-shadow: 0.3em 0.3em 0 0 var(--color), inset 0.3em 0.3em 0 0 var(--color);
+  box-shadow:
+    0.3em 0.3em 0 0 var(--color),
+    inset 0.3em 0.3em 0 0 var(--color);
 }
 .offset:hover,
 .offset:focus {
-  box-shadow: 0 0 0 0 var(--hover), inset 6em 3.5em 0 0 var(--hover);
+  box-shadow:
+    0 0 0 0 var(--hover),
+    inset 6em 3.5em 0 0 var(--hover);
 }
 .fill {
   --color: #e91e63;
