+++ src/shared/components/MainHeader/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@app/header-component",
+  "version": "0.1.0",
+  "private": true,
+  "keywords": [
+    "component",
+    "shareable"
+  ],
+  "main": "./src/index.ts"
+}
