+++ src/shared/components/MainHeader/src/MainHeader.tsx
@@ -1,5 +1,4 @@
-import React from "react";
-
+import "server-only";
 import styles from "./MainHeader.module.css";
 
 export const MainHeader = () => (
