+++ src/shared/components/Page/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@app/page-component",
+  "version": "0.1.0",
+  "private": true,
+  "keywords": [
+    "component",
+    "shareable"
+  ],
+  "main": "./src/index.ts"
+}
