+++ src/shared/utils/tsconfig.next.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../../../tsconfig.base.json",
+  "compilerOptions": { "outDir": "./dist", "rootDir": "." },
+  "references": []
+}
