+++ tsconfig.json
@@ -1,27 +1,26 @@
 {
+  "extends": "./tsconfig.base.json",
   "compilerOptions": {
-    "target": "ES2017",
-    "lib": ["dom", "dom.iterable", "esnext"],
-    "allowJs": true,
-    "skipLibCheck": true,
-    "strict": true,
-    "noEmit": true,
-    "esModuleInterop": true,
-    "module": "esnext",
-    "moduleResolution": "bundler",
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "jsx": "preserve",
-    "incremental": true,
-    "plugins": [
-      {
-        "name": "next"
-      }
-    ],
+    "outDir": "dist",
+    "baseUrl": "./",
     "paths": {
-      "@/*": ["./src/*"]
+      "@app/default-button": ["src/shared/components/DefaultButton/src"],
+      "@app/utils": ["src/shared/utils"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
-  "exclude": ["node_modules", ".next"]
+  "include": [
+    "**/*.ts",
+    "**/*.tsx",
+    "next-env.d.ts",
+    ".next/types/**/*.ts",
+    "next.config.js"
+  ],
+  "references": [
+    {
+      "path": "src/shared/components/DefaultButton/tsconfig.next.json"
+    },
+    {
+      "path": "src/shared/utils/tsconfig.next.json"
+    }
+  ]
 }
